broke settings page into 2 components, misc changes
