+++ src/containers/Settings/Settings.module.scss
@@ -9,11 +9,6 @@
   height: 15em;
 }
 
-.submitButton {
-  margin-top: $xxl-spacing;
-  text-align: center;
-}
-
 .submitError {
   margin-top: $base-spacing;
 }
