+++ src/form-components/index.js
@@ -6,8 +6,7 @@ import {
   RadioGroup,
   NumberInput,
   DatePicker,
-  Toggle,
-  Cropper
+  Toggle
 } from 'components';
 
 export const FormTextInput = FormInput(TextInput);
@@ -17,4 +16,3 @@ export const FormRadioGroup = FormInput(RadioGroup);
 export const FormNumberInput = FormInput(NumberInput);
 export const FormDatePicker = FormInput(DatePicker);
 export const FormToggle = FormInput(FormToggle);
-export const FormCropper = FormInput(FormCropper);
